Triangles and Meshes
Added intersection formulas for triangles. Created a RayTracingObject C# script that, when added to an object in unity, adds all of the meshes in the object to a buffer. The master script then takes each mesh and hands the triangles over to the shader via meshobject, vertex, and index buffers.